flask-dance/flask-dance-docker/requirements.txt
Flask
Flask-Dance
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
python-dotenv
psycopg2-binary
